Conflicts:
	app/code/Magento/ProductVideo/view/adminhtml/web/js/new-video-dialog.js

+++ app/code/Magento/Downloadable/Model/Link/DeleteHandler.php
@@ -1,16 +1,17 @@
 <?php
 /**
- * Copyright © 2015 Magento. All rights reserved.
+ * Copyright © 2016 Magento. All rights reserved.
  * See COPYING.txt for license details.
  */
 namespace Magento\Downloadable\Model\Link;
 
 use Magento\Downloadable\Api\LinkRepositoryInterface as LinkRepository;
+use Magento\Framework\EntityManager\Operation\ExtensionInterface;
 
 /**
  * Class DeleteHandler
  */
-class DeleteHandler
+class DeleteHandler implements ExtensionInterface
 {
     /**
      * @var LinkRepository
@@ -28,11 +29,15 @@ class DeleteHandler
     /**
      * @param string $entityType
      * @param object $entity
-     * @return object
+     * @param array $arguments
+     * @return \Magento\Catalog\Api\Data\ProductInterface|object
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
-    public function execute($entityType, $entity)
+    public function execute($entityType, $entity, $arguments = [])
     {
+        if ($entity->getTypeId() != \Magento\Downloadable\Model\Product\Type::TYPE_DOWNLOADABLE) {
+            return $entity;
+        }
         /** @var \Magento\Catalog\Api\Data\ProductInterface $entity */
         foreach ($this->linkRepository->getList($entity->getSku()) as $link) {
             $this->linkRepository->delete($link->getId());
